Renamed SideNav to Documentation, updated /grid path to /data
