+++ src/components/App/App.jsx
@@ -1,36 +1,36 @@
-import React, { useEffect, useState } from "react";
-import { Route, Switch } from "react-router-dom";
-import PrivateRoute from "../../hoc/PrivateRoute";
+import React, { useEffect, useState } from 'react'
+import { Route, Switch } from 'react-router-dom'
+import PrivateRoute from '../../hoc/PrivateRoute'
 
-import { PageView, initGA } from "../Tracking/Tracking";
-import Loading from "../Loading";
-import NavBar from "../NavBar";
-import Footer from "../Footer";
-import Profile from "../Profile";
-import Landing from "../Landing/index.js";
-import DocsPage from "../Docs/index";
-import GridPage from "../GridPage";
-import { useAuth0 } from "../../contexts";
+import { PageView, initGA } from '../Tracking/Tracking'
+import Loading from '../Loading'
+import NavBar from '../NavBar'
+import Footer from '../Footer'
+import Profile from '../Profile'
+import Landing from '../Landing/index.js'
+import DocsPage from '../Docs/index'
+import GridPage from '../GridPage'
+import { useAuth0 } from '../../contexts'
 
-import "./App.scss";
+import './App.scss'
 
 // fontawesome
-import initFontAwesome from "../../utils/initFontAwesome";
-initFontAwesome();
+import initFontAwesome from '../../utils/initFontAwesome'
+initFontAwesome()
 
 const App = () => {
-  const { loading } = useAuth0();
-  const [apiKey, setApiKey] = useState();
+  const { loading } = useAuth0()
+  const [apiKey, setApiKey] = useState()
   // const [list, setList] = useState(null)
 
   useEffect(() => {
     /*=== function that initializes Google Analytics ===*/
-    initGA(process.env.REACT_APP_GOOGLE_TRACKING_ID);
-    PageView();
-  });
+    initGA(process.env.REACT_APP_GOOGLE_TRACKING_ID)
+    PageView()
+  })
 
   if (loading) {
-    return <Loading />;
+    return <Loading />
   }
 
   return (
@@ -44,7 +44,7 @@ const App = () => {
             exact
             render={props => <GridPage {...props} apiKey={apiKey} />}
           /> */}
-        <Route exact path="/grid" component={GridPage} />
+        <Route exact path="/data" component={GridPage} />
         <PrivateRoute
           path="/profile"
           component={Profile}
@@ -55,7 +55,7 @@ const App = () => {
       </Switch>
       <Footer />
     </div>
-  );
-};
+  )
+}
 
-export default App;
+export default App
